model/EfficientNet-B4/requirements.txt
torch~=1.12.1+cu113
pillow~=10.4.0
torchvision~=0.13.1+cu113